Fix printf warnings in mbs plugin
git-svn-id: https://subversion.gsi.de/dabc/trunk@4956 bcbf6573-9a26-0410-9ebc-ce4ab7aade96

+++ plugins/mbs/src/CombinerModule.cxx
@@ -497,7 +497,7 @@ bool mbs::CombinerModule::BuildEvent()
       }
 
       if (!fOut.IsPlaceForEvent(subeventssize)) {
-         EOUT("Event size %u too big for buffer, skip event %u completely", subeventssize+ sizeof(mbs::EventHeader), buildevid);
+         EOUT("Event size %lu too big for buffer, skip event %u completely", (long unsigned) (subeventssize + sizeof(mbs::EventHeader)), buildevid);
       } else {
 
          if (copyMbsHdrId<0) {
